+++ Tests/Specs/DDNASmartAdServiceSpec.m
@@ -268,7 +268,7 @@ describe(@"interstitial ads", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@3];
         
         mockViewController = mock([UIViewController class]);
     });
@@ -495,7 +495,7 @@ describe(@"interstitial ads", ^{
         [(DDNAFakeSmartAdAgent *)fakeFactory.fakeSmartAdAgent closeAd];
 
         // session limit should be reached
-        expect([adService isInterstitialAdAvailable]).will.beTruthy();
+        expect([adService isInterstitialAdAvailable]).will.beFalsy();
         [adService showInterstitialAdFromRootViewController:mockViewController];
         [verifyCount(mockDelegate, times(3)) didOpenInterstitialAd];
 
@@ -539,7 +539,7 @@ describe(@"rewarded ads", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@3];
         
         mockViewController = mock([UIViewController class]);
     });
@@ -826,7 +826,7 @@ describe(@"respects minimum ad interval", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@3];
         
         mockViewController = mock([UIViewController class]);
     });
@@ -920,7 +920,7 @@ describe(@"respects adRequest flag", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@3];
         
         mockViewController = mock([UIViewController class]);
     });
@@ -985,7 +985,7 @@ describe(@"allowed to show interstitial", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@2];
         
         mockViewController = mock([UIViewController class]);
         
@@ -1062,7 +1062,7 @@ describe(@"allowed to show interstitial", ^{
         expect([adService isShowingInterstitialAd]).to.beTruthy();
         [(DDNAFakeSmartAdAgent *)fakeFactory.fakeSmartAdAgent closeAd];
         
-        expect([adService isInterstitialAdAvailable]).will.beTruthy();
+        expect([adService isInterstitialAdAvailable]).will.beFalsy();
         
         expect([adService isInterstitialAdAllowed]).to.beFalsy();
         
@@ -1118,7 +1118,7 @@ describe(@"allowed to show interstitial minimal time", ^{
                              }
                      };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@2];
         
         mockViewController = mock([UIViewController class]);
         
@@ -1228,7 +1228,7 @@ describe(@"allowed to show rewarded", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@2];
         
         mockViewController = mock([UIViewController class]);
         
@@ -1307,7 +1307,7 @@ describe(@"allowed to show rewarded", ^{
         expect([adService isShowingRewardedAd]).to.beTruthy();
         [(DDNAFakeSmartAdAgent *)fakeFactory.fakeSmartAdAgent closeAd];
         
-        expect([adService isRewardedAdAvailable]).will.beTruthy();
+        expect([adService isRewardedAdAvailable]).will.beFalsy();
         
         expect([adService isRewardedAdAllowed]).to.beFalsy();
         
@@ -1364,7 +1364,7 @@ describe(@"allowed to show rewarded minimal time", ^{
             }
         };
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@2];
         
         mockViewController = mock([UIViewController class]);
         
@@ -1453,7 +1453,7 @@ describe(@"respects adShowPoint and adShowSession for a session", ^{
         fakeFactory = [[DDNAFakeSmartAdFactory alloc] init];
         adService.factory = fakeFactory;
         
-        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] init];
+        fakeFactory.fakeSmartAdAgent = [[DDNAFakeSmartAdAgent alloc] initWithAdLimit:@2];
         
         mockViewController = mock([UIViewController class]);